Fixed scanner bug of LED turning off, but scan not being recognized
The scanner problem came from a UART buffer management problem on the
Imp where the Imp did not return partial buffers to the UART callback in
some cases such that the scan message was not parsed.
The fix creates a work-around and repeats the scan message 64 times in
order to create a full buffer. The Imp code only parses the first
message.
The long-term solution is for Imp to fix the UART buffer problem in Imp
OS.